Feedback to Player
Added player feedback. Added a simple starting menu that allows the player to choose a starting class. Added winand lose screens.

+++ MafiaSimulator/Assets/Scripts/InEditor_Scripts/ScriptableObjects/CardAsset/PlayerClass.cs
@@ -7,7 +7,7 @@ public class PlayerClass : ScriptableObject
     [Header("Player Starting Loadout")]
     private PlayerClass playerClass;
     public bool wasBouncer, wasAccountant, wasFarmer;
-    public float starting_Currency,starting_Athletics,Starting_Persuasion,Starting_Mobility;
+    public int starting_Currency,starting_Athletics,Starting_Persuasion,Starting_Mobility;
     public string PlayerClassName;
     public PlayerClass(PlayerClass playerClass)
     {
